backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pymongo==4.6.0
pydantic==2.5.0
python-dotenv==1.0.0
Pillow==10.1.0
google-cloud-aiplatform==1.38.1
google-cloud-vision==3.4.5
google-generativeai==0.3.2
langchain-core==0.1.0
langchain-huggingface==0.0.3
instagrapi==2.0.0
transformers==4.36.0
torch==2.1.1